refactor: renamed files to comply with coding standards
